requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
Babel==2.14.0
bcrypt==4.1.2
boto3==1.34.7
botocore==1.34.7
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.106.0
fastapi-admin==1.0.4
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
iso8601==1.1.0
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.0
mangum==0.17.0
MarkupSafe==2.1.4
pendulum==3.0.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
time-machine==2.13.0
tortoise-orm==0.20.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
